Vk: XFAIL image memory requirements test on a SwiftShader.
For some reason it wants me to allocate 16 bytes more. Why can't that be
stored somewhere else, I wonder?

Hm, and for this I implemented VK_KHR_driver_properties only to discover
that the info is not queryable if we run the tests with
KHR_get_physical_device_properties2 disabled. Sigh.